feat: Update Neovim plugins and colorscheme settings
This commit updates the versions of various Neovim plugins in the
lazy-lock.json file. It also modifies the colorscheme settings in the
colorscheme.lua file, adding a transparency option and changing the
style from "storm" to "night". The transparency settings for sidebars
and floats are also updated based on the new transparency variable.